pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plasma-reaction-builder"
version = "0.10.0"
description = "Evidence-first plasma state/reaction network builder"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "OpenAI"}
]
dependencies = [
  "requests>=2.31",
  "PyYAML>=6.0",
  "matplotlib>=3.8",
  "networkx>=3.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
plasma-rxn-builder = "plasma_reaction_builder.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
plasma_reaction_builder = ["data/*.yaml"]
